--- src/router/routes/admin-routes.tsx
@@ -1,9 +1,0 @@
-
-import AdminEmailCampaigns from "@/pages/AdminEmailCampaigns";
-
-export const adminRoutes = [
-  {
-    path: "admin/email-campaigns",
-    element: <AdminEmailCampaigns />,
-  },
-];
